parent_index in ConsumeNext for ExecNodes should be size_t, not int64_t
Summary: I made it int64_t before but size_t is more correct because it's always non-negative.

Test Plan: existing

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D1263

GitOrigin-RevId: ed430e0d0069fda486151a876e64376d81f57a51

+++ src/carnot/exec/blocking_agg_node.cc
@@ -128,7 +128,7 @@ Status BlockingAggNode::OpenImpl(ExecState* exec_state) {
   return Status::OK();
 }
 
-Status BlockingAggNode::ConsumeNextImpl(ExecState* exec_state, const RowBatch& rb, int64_t) {
+Status BlockingAggNode::ConsumeNextImpl(ExecState* exec_state, const RowBatch& rb, size_t) {
   if (HasNoGroups()) {
     return AggregateGroupByNone(exec_state, rb);
   }
