+++ src/carnot/exec/memory_sink_node.h
@@ -25,7 +25,7 @@ class MemorySinkNode : public SinkNode {
   Status OpenImpl(ExecState* exec_state) override;
   Status CloseImpl(ExecState* exec_state) override;
   Status ConsumeNextImpl(ExecState* exec_state, const table_store::schema::RowBatch& rb,
-                         int64_t parent_index) override;
+                         size_t parent_index) override;
 
  private:
   std::unique_ptr<plan::MemorySinkOperator> plan_node_;
